move KVPairsAttrs to shared.templ
