--- internal/features/common/components/shared_partials.go
@@ -1,14 +1,0 @@
-package components
-
-import "github.com/a-h/templ"
-
-func KVPairsAttrs(kvPairs ...string) templ.Attributes {
-	if len(kvPairs)%2 != 0 {
-		panic("kvPairs must be a multiple of 2")
-	}
-	attrs := templ.Attributes{}
-	for i := 0; i < len(kvPairs); i += 2 {
-		attrs[kvPairs[i]] = kvPairs[i+1]
-	}
-	return attrs
-}
