+++ internal/features/common/components/shared_templ.go
@@ -10,6 +10,17 @@ import templruntime "github.com/a-h/templ/runtime"
 
 import "fmt"
 
+func KVPairsAttrs(kvPairs ...string) templ.Attributes {
+	if len(kvPairs)%2 != 0 {
+		panic("kvPairs must be a multiple of 2")
+	}
+	attrs := templ.Attributes{}
+	for i := 0; i < len(kvPairs); i += 2 {
+		attrs[kvPairs[i]] = kvPairs[i+1]
+	}
+	return attrs
+}
+
 func Icon(icon string, attrs ...string) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
@@ -38,7 +49,7 @@ func Icon(icon string, attrs ...string) templ.Component {
 		var templ_7745c5c3_Var2 string
 		templ_7745c5c3_Var2, templ_7745c5c3_Err = templ.JoinStringErrs(icon)
 		if templ_7745c5c3_Err != nil {
-			return templ.Error{Err: templ_7745c5c3_Err, FileName: `internal/features/common/components/shared.templ`, Line: 6, Col: 26}
+			return templ.Error{Err: templ_7745c5c3_Err, FileName: `internal/features/common/components/shared.templ`, Line: 17, Col: 26}
 		}
 		_, templ_7745c5c3_Err = templ_7745c5c3_Buffer.WriteString(templ.EscapeString(templ_7745c5c3_Var2))
 		if templ_7745c5c3_Err != nil {
@@ -88,7 +99,7 @@ func SseIndicator(signalName string) templ.Component {
 		var templ_7745c5c3_Var4 string
 		templ_7745c5c3_Var4, templ_7745c5c3_Err = templ.JoinStringErrs(fmt.Sprintf("{'loading ml-4': $%s}", signalName))
 		if templ_7745c5c3_Err != nil {
-			return templ.Error{Err: templ_7745c5c3_Err, FileName: `internal/features/common/components/shared.templ`, Line: 10, Col: 101}
+			return templ.Error{Err: templ_7745c5c3_Err, FileName: `internal/features/common/components/shared.templ`, Line: 21, Col: 101}
 		}
 		_, templ_7745c5c3_Err = templ_7745c5c3_Buffer.WriteString(templ.EscapeString(templ_7745c5c3_Var4))
 		if templ_7745c5c3_Err != nil {
